requirements.txt
openenv-core>=0.2.1
fastapi>=0.115.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
openai>=1.0.0
datasets>=2.0.0
